users and user got the same structur as company
