+++ priv/react_app/src/App/App.js
@@ -23,7 +23,7 @@ import Deadline from '../Screens/Admin/Deadline';
 import Roles from '../Screens/Admin/Roles';
 import Role from '../Screens/Admin/Role';
 import Users from '../Screens/Admin/Users';
-import User from '../Screens/Admin/User';
+import { UserEdit, UserShow } from '../Screens/Admin/User';
 import CurrentUser from '../Screens/CurrentUser';
 import Companies from '../Screens/Admin/Companies';
 import { CompanyNew } from '../Screens/Admin/Company';
@@ -64,7 +64,8 @@ const routes = (
     <PrivateRoute exact path="/admin/deadlines" component={Deadlines} />
     <PrivateRoute path="/admin/deadlines/:id" component={Deadline} />
     <PrivateRoute exact path="/admin/users" component={Users} />
-    <PrivateRoute path="/admin/users/:id" component={User} />
+    <PrivateRoute exact path="/admin/users/:id/edit" component={UserEdit} />
+    <PrivateRoute path="/admin/users/:id" component={UserShow} />
     <PrivateRoute exact path="/admin/roles" component={Roles} />
     <PrivateRoute path="/admin/roles/:id" component={Role} />
     <Route path="/login" component={Login} />
