+++ priv/react_app/src/Screens/Admin/User/index.js
@@ -1,30 +1,4 @@
-import { connect } from 'react-redux';
-import { denormalize } from 'normalizr';
-import { Actions } from '../../../Store';
-import Schema from '../../../Store/normalizr/schema';
-import User from './User';
+import UserEdit from './UserEdit';
+import UserShow from './UserShow';
 
-const mapStateToProps = (state, props) => {
-  const userId = props.match.params.id;
-  const user = state.entities.users[userId] || {};
-
-  const { roles } = denormalize(
-    { roles: user.roles },
-    Schema.userSchema(),
-    state.entities
-  );
-
-  return { id: userId, user: { ...user, roles } };
-};
-
-const mapDispatchToProps = {
-  getUser: Actions.users.getUser,
-  updateUser: Actions.users.updateUser
-};
-
-const stateful = connect(
-  mapStateToProps,
-  mapDispatchToProps
-);
-
-export default stateful(User);
+export { UserEdit, UserShow };
