+++ priv/react_app/src/Screens/Admin/Users/Users.js
@@ -2,46 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import Table from 'antd/lib/table';
-import Button from 'antd/lib/button';
 import Divider from 'antd/lib/divider';
 import Popconfirm from 'antd/lib/popconfirm';
 import InvisibleLink from '../../../Components/InvisibleLink';
 import LoadingSpinner from '../../../Components/LoadingSpinner';
 import HtmlTitle from '../../../Components/HtmlTitle';
-
-const userColumns = [
-  {
-    title: 'Email',
-    dataIndex: 'email',
-    key: 'email',
-    render: (email, { id }) => (
-      <InvisibleLink to={`/admin/users/${id}`}>{email}</InvisibleLink>
-    )
-  },
-  {
-    title: 'First Name',
-    dataIndex: 'firstName',
-    key: 'firstName'
-  },
-  {
-    title: 'Last Name',
-    dataIndex: 'lastName',
-    key: 'lastName'
-  },
-  {
-    title: 'Action',
-    key: 'action',
-    render: user => (
-      <span>
-        <InvisibleLink to={`/admin/users/${user.id}`}>Show</InvisibleLink>
-        <Divider type="vertical" />
-        <InvisibleLink to="#user-edit">Edit</InvisibleLink>
-        <Divider type="vertical" />
-        <InvisibleLink to="#user-delete">Delete</InvisibleLink>
-      </span>
-    )
-  }
-];
 
 /**
  * Responsible for rendering a list of users
@@ -52,47 +17,50 @@ class Users extends Component {
     getAllUsers();
   }
 
-  userColumns = () => [
-    {
-      title: 'Email',
-      dataIndex: 'email',
-      key: 'email',
-      render: (email, { id }) => (
-        <InvisibleLink to={`/users/${id}`}>{email}</InvisibleLink>
-      )
-    },
-    {
-      title: 'First Name',
-      dataIndex: 'firstName',
-      key: 'firstName'
-    },
-    {
-      title: 'Last Name',
-      dataIndex: 'lastName',
-      key: 'lastName'
-    },
-    {
-      title: 'Action',
-      key: 'action',
-      render: user => (
-        <span>
-          <InvisibleLink to={`/users/${user.id}`}>Show</InvisibleLink>
-          <Divider type="vertical" />
-          <InvisibleLink to={`/users/${user.id}#edit`}>Edit</InvisibleLink>
-          <Divider type="vertical" />
-          <Popconfirm
-            title="Sure to delete?"
-            onConfirm={() => this.props.deleteUser(user.id)}
-          >
-            <span style={{ color: '#ff4d4f', cursor: 'pointer' }}>Delete</span>
-          </Popconfirm>
-        </span>
-      )
-    }
-  ];
-
   renderUsers() {
     const { users } = this.props;
+    const userColumns = [
+      {
+        title: 'Email',
+        dataIndex: 'email',
+        key: 'email',
+        render: (email, { id }) => (
+          <InvisibleLink to={`/admin/users/${id}`}>{email}</InvisibleLink>
+        )
+      },
+      {
+        title: 'First Name',
+        dataIndex: 'firstName',
+        key: 'firstName'
+      },
+      {
+        title: 'Last Name',
+        dataIndex: 'lastName',
+        key: 'lastName'
+      },
+      {
+        title: 'Action',
+        key: 'action',
+        render: user => (
+          <span>
+            <InvisibleLink to={`/admin/users/${user.id}`}>Show</InvisibleLink>
+            <Divider type="vertical" />
+            <InvisibleLink to={`/admin/users/${user.id}/edit`}>
+              Edit
+            </InvisibleLink>
+            <Divider type="vertical" />
+            <Popconfirm
+              title="Sure to delete?"
+              onConfirm={() => this.props.deleteUser(user.id)}
+            >
+              <span style={{ color: '#ff4d4f', cursor: 'pointer' }}>
+                Delete
+              </span>
+            </Popconfirm>
+          </span>
+        )
+      }
+    ];
 
     return (
       <div>
@@ -101,16 +69,12 @@ class Users extends Component {
         <h1>Users</h1>
 
         <Table
-          columns={this.userColumns()}
+          columns={userColumns}
           dataSource={Object.keys(users).map(i => ({
             ...users[i],
             key: i
           }))}
         />
-
-        <Button onClick={() => null} type="primary">
-          New user
-        </Button>
       </div>
     );
   }
