--- priv/react_app/src/Screens/Users/index.js
@@ -1,21 +1,0 @@
-import { connect } from 'react-redux';
-import Users from './Users';
-import { Actions } from '../../Store';
-import type { State } from '../../Store/reducers';
-
-const mapStateToProps = (state: State) => ({
-  users: state.entities.users,
-  fetching: state.api.users.fetching
-});
-
-const mapDispatchToProps = {
-  getAllUsers: Actions.users.getAllUsers,
-  deleteUser: Actions.users.destroyUser
-};
-
-const stateful = connect(
-  mapStateToProps,
-  mapDispatchToProps
-);
-
-export default stateful(Users);
